python/requirements.txt
requests
python-dotenv
pika